Version 4 Authentication Complete

+++ app.py
@@ -1,5 +1,18 @@
-from flask import Flask, render_template, request, redirect, flash
+from flask import (
+    Flask,
+    render_template,
+    request,
+    redirect,
+    flash,
+    session
+)
+
 from flask_sqlalchemy import SQLAlchemy
+
+from werkzeug.security import (
+    generate_password_hash,
+    check_password_hash
+)
 
 app = Flask(__name__)
 app.secret_key = "expense_tracker_secret"
@@ -61,10 +74,80 @@ class Expense(db.Model):
         nullable=False
     )
 
-user_id = db.Column(
-    db.Integer,
-    db.ForeignKey('user.id')
-)
+@app.route('/register', methods=['GET', 'POST'])
+def register():
+
+    if request.method == 'POST':
+
+        username = request.form['username']
+        email = request.form['email']
+        password = request.form['password']
+        confirm_password = request.form['confirm_password']
+
+        if password != confirm_password:
+            flash("Passwords do not match")
+            return redirect('/register')
+
+        existing_user = User.query.filter_by(
+            email=email
+        ).first()
+
+        if existing_user:
+            flash("Email already exists")
+            return redirect('/register')
+
+        hashed_password = generate_password_hash(
+            password
+        )
+
+        user = User(
+            username=username,
+            email=email,
+            password=hashed_password
+        )
+
+        db.session.add(user)
+        db.session.commit()
+
+        flash("Registration Successful")
+
+        return redirect('/login')
+
+    return render_template('register.html')
+
+@app.route('/login', methods=['GET', 'POST'])
+def login():
+
+    if request.method == 'POST':
+
+        email = request.form['email']
+        password = request.form['password']
+
+        user = User.query.filter_by(
+            email=email
+        ).first()
+
+        if user and check_password_hash(
+            user.password,
+            password
+        ):
+
+            session['user_id'] = user.id
+            session['username'] = user.username
+
+            return redirect('/')
+
+        flash("Invalid Email or Password")
+
+    return render_template('login.html')
+
+@app.route('/logout')
+def logout():
+
+    session.clear()
+
+    return redirect('/login')
+
 
 with app.app_context():
     db.create_all()
@@ -73,17 +156,22 @@ with app.app_context():
 @app.route('/', methods=['GET', 'POST'])
 def home():
 
+    if 'user_id' not in session:
+        return redirect('/login')
+
     if request.method == 'POST':
 
         name = request.form['name']
         amount = request.form['amount']
         category = request.form['category']
 
+        
         expense = Expense(
             name=name,
             amount=int(amount),
-            category=category
-        )
+            category=category,
+            user_id=session['user_id']
+    )
 
         db.session.add(expense)
         db.session.commit()
@@ -92,7 +180,9 @@ def home():
 
         return redirect('/')
 
-    expenses = Expense.query.order_by(
+    expenses = Expense.query.filter_by(
+        user_id=session['user_id']
+    ).order_by(
         Expense.id.desc()
     ).all()
 
@@ -163,7 +253,13 @@ def home():
 @app.route('/update/<int:id>', methods=['POST'])
 def update(id):
 
-    expense = Expense.query.get_or_404(id)
+    if 'user_id' not in session:
+        return redirect('/login')
+
+    expense = Expense.query.filter_by(
+        id=id,
+        user_id=session['user_id']
+    ).first_or_404()
 
     expense.name = request.form['name']
     expense.amount = int(
@@ -181,7 +277,13 @@ def update(id):
 @app.route('/delete/<int:id>')
 def delete(id):
 
-    expense = Expense.query.get_or_404(id)
+    if 'user_id' not in session:
+        return redirect('/login')
+
+    expense = Expense.query.filter_by(
+        id=id,
+        user_id=session['user_id']
+    ).first_or_404()
 
     db.session.delete(expense)
     db.session.commit()
